# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управління транспортними засобами (автомобілі, бренди, категорії тощо).",
        "title": "GoVehicleApi",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/body-style/refresh": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "body-style"
                ],
                "summary": "Синхронізувати стилі кузова з Autoria",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/body-style/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "body-style"
                ],
                "summary": "Отримати стиль кузова за ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стилю кузова",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body_style.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "body-style"
                ],
                "summary": "Оновити стиль кузова",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стилю кузова",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Нові дані",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body_style.BodyStylePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body_style.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "body-style"
                ],
                "summary": "Видалити стиль кузова",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стилю кузова",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/brand/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brand"
                ],
                "summary": "Список брендів",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brand.GetAllResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/brand/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brand"
                ],
                "summary": "Отримати бренд за ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бренду",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brand.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brand"
                ],
                "summary": "Оновити бренд",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бренду",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Нові дані",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/brand.BrandPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/brand.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brand"
                ],
                "summary": "Видалити бренд",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бренду",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "body_style.BodyStylePayload": {
            "type": "object",
            "properties": {
                "BodyStyle": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "value": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "body_style.GetResponse": {
            "type": "object",
            "properties": {
                "bodyStyle": {
                    "$ref": "#/definitions/models.BodyStyle"
                },
                "response": {
                    "$ref": "#/definitions/response.Response"
                }
            }
        },
...
        "body_style.UpdateResponse": {
            "type": "object",
            "properties": {
                "bodyStyle": {
                    "$ref": "#/definitions/models.BodyStyle"
                },
                "response": {
                    "$ref": "#/definitions/response.Response"
                }
            }
        },
        "brand.BrandPayload": {
            "type": "object",
            "properties": {
                "Brand": {
                    "type": "object",
                    "properties": {
                        "category_id": {
                            "type": "integer"
                        },
                        "cnt": {
                            "type": "integer"
                        },
                        "country_id": {
                            "type": "integer"
                        },
                        "eng": {
                            "type": "string"
                        },
                        "marka_id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "slang": {
                            "type": "string"
                        },
                        "value": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "brand.GetAllResponse": {
            "type": "object",
            "properties": {
                "brands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Brand"
                    }
                },
                "response": {
                    "$ref": "#/definitions/response.Response"
                }
            }
        },
        "brand.GetResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/models.Brand"
                },
                "response": {
                    "$ref": "#/definitions/response.Response"
                }
            }
        },
...
        "brand.UpdateResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/models.Brand"
                },
                "response": {
                    "$ref": "#/definitions/response.Response"
                }
            }
        },
...
        "models.BodyStyle": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "models.Brand": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "cnt": {
                    "type": "integer"
                },
                "country_id": {
                    "type": "integer"
                },
                "eng": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "marka_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slang": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введіть токен у форматі: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}